fixed new dependency issues with react-router
