+++ src/App.tsx
@@ -9,10 +9,10 @@ export default () => {
     return <BrowserRouter>
         <NavController />
         <Switch>
-            <Route exact path="/" ><HomePage /></Route>
-            <Route exact path="/e-commerce/"><EcommercePage /></Route>
+            <Route path="/e-commerce/"><EcommercePage /></Route>
             <Route path="/transaction/:id/" ><TransactionPage /></Route>
             <Route path="/transaction/"><TransactionListPage /></Route>
+            <Route path="/" ><HomePage /></Route>
         </Switch>
     </BrowserRouter>;
 };
