+++ src/controllers/MetamaskErrors.tsx
@@ -1,25 +1,6 @@
 import useProviderError from "hooks/useProviderError";
-import React, { useState } from "react";
-import { MetamaskErrorSeverity } from "types/enums";
-
-
-
-const MetamaskErrorView = ({ severity, name, description }: { severity: MetamaskErrorSeverity, name: MetamaskErrorName, description: string }) => {
-    const classSeverity = severity === MetamaskErrorSeverity.BLOCKING ? " error" : (severity === MetamaskErrorSeverity.UNBLOCKING ? " warning" : " success");
-    const iconSeverity = severity === MetamaskErrorSeverity.BLOCKING ? "✗" : (severity === MetamaskErrorSeverity.UNBLOCKING ? "⚠" : "✓");
-    const [hidden, setHidden] = useState<boolean>(false);
-
-    const toggle = () => setHidden(!hidden);
-
-    return <span className={"provider-error" + classSeverity} onClick={toggle}>
-        {iconSeverity}
-        <div className={"error-overlay"} onClick={toggle}>
-            <span>{name}</span>
-            <p>{description}</p>
-        </div>
-    </span>;
-};
-
+import React from "react";
+import { MetamaskErrorView } from "views";
 
 const MetamaskErrorController = () => {
     const { severity, name, description } = useProviderError();
