+++ src/resources/svg.tsx
@@ -70,6 +70,10 @@ export const Loading = (props: IProps) => <SvgIcon {...props} viewBox="0 0 100 1
     </g>
 </SvgIcon>;
 
+export const CheckMark = (props: IProps) => <SvgIcon {...props}><path d="M0 0h24v24H0V0z" fill="none"/><path d="M9 16.2L4.8 12l-1.4 1.4L9 19 21 7l-1.4-1.4L9 16.2z"/></SvgIcon>;
+export const Cross = (props:  IProps) => <SvgIcon {...props}><path d="M0 0h24v24H0V0z" fill="none"/><path d="M19 6.41L17.59 5 12 10.59 6.41 5 5 6.41 10.59 12 5 17.59 6.41 19 12 13.41 17.59 19 19 17.59 13.41 12 19 6.41z"/></SvgIcon>;
+export const Warning = (props: IProps) => <SvgIcon {...props}><path d="M0 0h24v24H0V0z" fill="none"/><path d="M12 5.99L19.53 19H4.47L12 5.99M12 2L1 21h22L12 2zm1 14h-2v2h2v-2zm0-6h-2v4h2v-4z"/></SvgIcon>;
+
 export const FTMIcon = (props: IProps) => <SvgIcon {...props} viewBox='0 0 32 32'>
     <defs xmlns="http://www.w3.org/2000/svg">
         <mask id="mask" x="10" y="6" width="93.1" height="20" maskUnits="userSpaceOnUse"><g id="a">
